@@ -1,5 +1,7 @@
 <?php
-namespace NetDriver\Exception;
+declare(strict_types=1);
+
+namespace Stk2k\NetDriver\Exception;
 
 interface NetDriverExceptionInterface
 {
