@@ -1,8 +1,10 @@
 <?php
-namespace NetDriver\Http;
+declare(strict_types=1);
 
-use NetDriver\Enum\EnumHttpMethod;
-use NetDriver\NetDriverInterface;
+namespace Stk2k\NetDriver\Http;
+
+use Stk2k\NetDriver\Enum\EnumHttpMethod;
+use Stk2k\NetDriver\NetDriverInterface;
 
 class HttpPostRequest extends HttpRequest
 {
@@ -17,7 +19,7 @@ class HttpPostRequest extends HttpRequest
      * @param array $post_fields
      * @param array $options
      */
-    public function __construct($driver, $url, array $post_fields, array $options = [])
+    public function __construct(NetDriverInterface $driver, string $url, array $post_fields, array $options = [])
     {
         parent::__construct($driver, EnumHttpMethod::POST, $url, $options);
 
@@ -27,8 +29,8 @@ class HttpPostRequest extends HttpRequest
     /**
      * @return string
      */
-    public function getPostFields()
+    public function getPostFields() : string
     {
-        return http_build_query($this->post_fields, "", "&");
+        return http_build_query($this->post_fields);
     }
 }
