@@ -1,50 +1,52 @@
 <?php
-namespace NetDriver;
+declare(strict_types=1);
+
+namespace Stk2k\NetDriver;
 
 use Psr\Log\LoggerInterface;
 
-use NetDriver\Http\HttpRequest;
-use NetDriver\Http\HttpResponse;
-use NetDriver\Exception\NetDriverException;
-use NetDriver\Exception\TimeoutException;
-use NetDriver\Exception\DeflateException;
+use Stk2k\NetDriver\Http\HttpRequest;
+use Stk2k\NetDriver\Http\HttpResponse;
+use Stk2k\NetDriver\Exception\NetDriverException;
+use Stk2k\NetDriver\Exception\TimeoutException;
+use Stk2k\NetDriver\Exception\DeflateException;
 
 interface NetDriverInterface
 {
     /**
      * Set verbose flag
      *
-     * @param $verbose
+     * @param bool $verbose
      */
-    public function setVerbose($verbose = true);
+    public function setVerbose(bool $verbose = true);
 
     /**
      * Get verbose flag
      *
      * @return bool
      */
-    public function getVerbose();
+    public function getVerbose() : bool;
 
     /**
      * Set user agent
      *
      * @param $user_agent
      */
-    public function setUserAgent($user_agent);
+    public function setUserAgent(string $user_agent);
 
     /**
      * Get user agent
      *
      * @return string
      */
-    public function getUserAgent();
+    public function getUserAgent() : string;
 
     /**
      * Create new handle
      *
      * @return NetDriverHandleInterface
      */
-    public function newHandle();
+    public function newHandle() : NetDriverHandleInterface;
 
     /**
      * Send HTTP request
@@ -58,7 +60,7 @@ interface NetDriverInterface
      * @throws TimeoutException
      * @throws DeflateException
      */
-    public function sendRequest(NetDriverHandleInterface $handle, HttpRequest $request);
+    public function sendRequest(NetDriverHandleInterface $handle, HttpRequest $request) : HttpResponse;
 
     /**
      * Listen event
@@ -66,7 +68,7 @@ interface NetDriverInterface
      * @param string $event
      * @param callable $listener
      */
-    public function listen($event, $listener);
+    public function listen(string $event, callable $listener);
 
     /**
      * Set logger
